TLKAR2-5: run kartograafr under Docker. Checkpoint
