+++ Dockerfile.md
@@ -1,0 +1,6 @@
+Used official python 2.7 image from Dockerhub.
+
+https://hub.docker.com/_/python/
+
+To run and build for development see runDocker.sh.
+
